+++ zcord/models/component/__init__.py
@@ -1,5 +1,6 @@
 from .action_row import ActionRow
 from .base import Component
 from .button import Button
+from .select_menu import StringSelect
 
-__all__ = ["ActionRow", "Button", "Component"]
+__all__ = ["ActionRow", "Button", "Component", "StringSelect"]