+++ zcord/models/component/action_row.py
@@ -9,12 +9,21 @@ from zcord.models.component.base import Component
 
 if TYPE_CHECKING:
     from zcord.models.component.button import Button
+    from zcord.types import SelectMenu
 
 
 @dataclass(frozen=True, slots=True)
 class ActionRow(Component):
+    """
+    Represent an action row that holds max to 5 buttons or a select menu.
+
+    Attributes:
+        components:
+            A list of components inside the action row.
+    """
+
     type: ComponentType = ComponentType.ACTION_ROW
-    components: list[Button] | Component | MISSING = MISSING
+    components: list[Button] | SelectMenu | MISSING = MISSING
 
     _transforms: ClassVar[dict] = {
         "type": ComponentType,