+++ zcord/types.py
@@ -1,0 +1,9 @@
+"""
+Classes in this module are used for interfacing with the Discord API.
+
+These hold no attributes or methods and are only used for type hinting.
+"""
+
+
+class SelectMenu:
+    """A select menu component type."""