+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project uses mixed Calendar versioning: YYYY.feature.patch(.tag)
 ### Added:
   - Models:
     - `SharedClientTheme`, `BaseThemeType` enums and shared client theme builder.
+    - `Component`s
+      - `Button`
+      - `ActionRow`
+      - `StringSelect`
   - `Poll` builder.
 
 ## 2026.0.2 - 2026/07/26